Fixed some bugs

+++ 2.1-3-searching.cc
@@ -3,7 +3,7 @@
 // at the start of each iteration of the for loop, variable result holds true state
 mvalue search(array &A, int v) {
   mvalue result;
-  for (int i = 1; i <= A.length; ++i)
+  for (size_t i = 1; i <= A.length; ++i)
     if (A[i] == v) {
       result = i;
       break;
