+++ 2.3-7-two_integer_sum.cc
@@ -3,7 +3,7 @@
 bool sum_exists(array &S, int x) {
   merge_sort(S);
   bool exists = false;
-  for (int i = 1; i <= S.length; ++i)
+  for (size_t i = 1; i <= S.length; ++i)
     if (!bsearch(S, x - S[i]).nil) {
       exists = true;
       break;
